requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.7
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docxtpl==0.20.2
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.14
drf-yasg2==1.19.4
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.20.0
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.8.0
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.25.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lxml==6.0.2
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib==3.8.0
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.7.0
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.4.0
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.4.1
psutil==7.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.2
pytest-timeout==2.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.3
pyzmq==26.4.0
qrcode==8.2
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.29.0
ruamel.yaml==0.19.1
scikit-learn==1.3.0
scipy==1.15.3
seaborn==0.13.0
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.5
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
yarl==1.22.0
